Move CopyOnAccessStateInternals to runners/direct
