+++ sdks/java/core/src/main/java/org/apache/beam/sdk/util/state/InMemoryStateInternals.java
@@ -56,7 +56,11 @@ public class InMemoryStateInternals<K> implements StateInternals<K> {
     return key;
   }
 
-  interface InMemoryState<T extends InMemoryState<T>> {
+  /**
+   * Interface common to all in-memory state cells. Includes ability to see whether a cell has been
+   * cleared and the ability to create a clone of the contents.
+   */
+  public interface InMemoryState<T extends InMemoryState<T>> {
     boolean isCleared();
     T copy();
   }
@@ -94,11 +98,11 @@ public class InMemoryStateInternals<K> implements StateInternals<K> {
   /**
    * A {@link StateBinder} that returns In Memory {@link State} objects.
    */
-  static class InMemoryStateBinder<K> implements StateBinder<K> {
+  public static class InMemoryStateBinder<K> implements StateBinder<K> {
     private final K key;
     private final StateContext<?> c;
 
-    InMemoryStateBinder(K key, StateContext<?> c) {
+    public InMemoryStateBinder(K key, StateContext<?> c) {
       this.key = key;
       this.c = c;
     }
@@ -150,7 +154,11 @@ public class InMemoryStateInternals<K> implements StateInternals<K> {
     }
   }
 
-  static final class InMemoryValue<T> implements ValueState<T>, InMemoryState<InMemoryValue<T>> {
+  /**
+   * An {@link InMemoryState} implementation of {@link ValueState}.
+   */
+  public static final class InMemoryValue<T>
+      implements ValueState<T>, InMemoryState<InMemoryValue<T>> {
     private boolean isCleared = true;
     private T value = null;
 
@@ -194,7 +202,10 @@ public class InMemoryStateInternals<K> implements StateInternals<K> {
     }
   }
 
-  static final class InMemoryWatermarkHold<W extends BoundedWindow>
+  /**
+   * An {@link InMemoryState} implementation of {@link WatermarkHoldState}.
+   */
+  public static final class InMemoryWatermarkHold<W extends BoundedWindow>
       implements WatermarkHoldState<W>, InMemoryState<InMemoryWatermarkHold<W>> {
 
     private final OutputTimeFn<? super W> outputTimeFn;
@@ -267,7 +278,10 @@ public class InMemoryStateInternals<K> implements StateInternals<K> {
     }
   }
 
-  static final class InMemoryCombiningValue<K, InputT, AccumT, OutputT>
+  /**
+   * An {@link InMemoryState} implementation of {@link AccumulatorCombiningState}.
+   */
+  public static final class InMemoryCombiningValue<K, InputT, AccumT, OutputT>
       implements AccumulatorCombiningState<InputT, AccumT, OutputT>,
           InMemoryState<InMemoryCombiningValue<K, InputT, AccumT, OutputT>> {
     private final K key;
@@ -275,7 +289,7 @@ public class InMemoryStateInternals<K> implements StateInternals<K> {
     private final KeyedCombineFn<? super K, InputT, AccumT, OutputT> combineFn;
     private AccumT accum;
 
-    InMemoryCombiningValue(
+    public InMemoryCombiningValue(
         K key, KeyedCombineFn<? super K, InputT, AccumT, OutputT> combineFn) {
       this.key = key;
       this.combineFn = combineFn;
@@ -353,7 +367,10 @@ public class InMemoryStateInternals<K> implements StateInternals<K> {
     }
   }
 
-  static final class InMemoryBag<T> implements BagState<T>, InMemoryState<InMemoryBag<T>> {
+  /**
+   * An {@link InMemoryState} implementation of {@link BagState}.
+   */
+  public static final class InMemoryBag<T> implements BagState<T>, InMemoryState<InMemoryBag<T>> {
     private List<T> contents = new ArrayList<>();
 
     @Override
